Inter first steps
Made a start with the modelling of Inter transport. Trucks for inter transport are now initialized, and updated during the simulation. When an inter truck reaches its destination it picks up sorted RC from the depot. The functionality is not yet complete. Currently no distinction is made between crossdocks and depots, so this needs to be further refined. Some data cleaning needs to be added to convert the VAR names for depots and crossdocks to those in the simulation.